refactor: Move SchemaValidator implementations into DB modules
